requirements.txt
fastapi
uvicorn[standard]
pillow
opencv-python-headless
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose
